Adjusting models to match the data structure in the jsonfiles.

+++ Models/IndustryExperience.cs
@@ -3,7 +3,7 @@ namespace ProfileApi.Models;
 public class IndustryExperience
 {
     public string Skill { get; set; } = string.Empty;
-    public int Years { get; set; }
+    public double Years { get; set; }
     public string Level { get; set; } = string.Empty;
     public string Domain { get; set; } = string.Empty;
     public string DomainIcon { get; set; } = string.Empty;
